+++ ProjectFuji/GeneralGrid.h
@@ -31,7 +31,7 @@ public:
 		\param[in] range		Range of the x/y grid.
 		\param[in] stepSize		Size of the spacing between lines in the x/y grid.
 	*/
-	GeneralGrid(float range, float stepSize);
+	GeneralGrid(int range, int stepSize);
 
 	//! Default destructor.
 	~GeneralGrid();
@@ -48,8 +48,8 @@ private:
 
 	int numLines;	//!< Number of lines to draw for the x/z grid
 
-	float range;	//!< Range of the x/z grid
-	float stepSize;	//!< Number of steps (spacing) between individual lines of the x/z grid
+	int range;	//!< Range of the x/z grid
+	int stepSize;	//!< Number of steps (spacing) between individual lines of the x/z grid
 
 
 };
